+++ src/index-standalone.ts
@@ -1,0 +1,2 @@
+export { setOptions } from "@zappar/zappar-threejs/lib/index-standalone";
+export * from "./index";