+++ src/index.ts
@@ -3,7 +3,7 @@ import VERSION from "./version";
 // eslint-disable-next-line no-console
 console.log(`Zappar for React Three v${VERSION}`);
 
-export { browserIncompatible, Pipeline, LogLevel, setLogLevel } from "@zappar/zappar-threejs";
+export { browserIncompatible, Pipeline, LogLevel, setLogLevel, SequenceSource } from "@zappar/zappar-threejs";
 export { Types } from "./spec";
 export { default as FaceBufferGeometry } from "./Components/FaceBufferGeometry";
 export { default as FaceLandmark } from "./Components/trackers/FaceLandmarkGroup";
@@ -15,4 +15,5 @@ export { default as ZapparCamera } from "./Components/ZapparCamera";
 export { default as ZapparCanvas } from "./Components/ZapparCanvas";
 export { default as Loader } from "./Components/util/Loader";
 export { default as BrowserCompatibility } from "./Components/util/Compatibility";
+export { default as TargetImagePreviewMesh } from "./Components/targetImagePreviewMesh";
 // export { default as CameraEnvironmentMap } from "./Components/CameraEnvironmentMap";