+++ tests/camera-environment-map-global-cam-prop.ts
@@ -11,7 +11,7 @@ const url = "https://0.0.0.0:8082/pages/jest/module/camera-environment-map-globa
 describe("face tracking", () => {
   it("console logs", async () => {
     const page = await browser.newPage();
-    page.goto(url);
+    page.goto(url, { timeout: 0 });
     await util.expectLogs({
       expected: [
         /^Zappar for React Three v/,
@@ -24,7 +24,7 @@ describe("face tracking", () => {
         "[Zappar] INFO identity for license check: 0.0.0.0",
       ],
       page,
-      timeoutMs: 30000,
+      timeoutMs: 60000,
     });
 
     // Wait 1s for texture to load