+++ tests/jest/module/_template0_simple-image.tsx
@@ -4,14 +4,17 @@
 import React, { useMemo, useState } from "react";
 import { render } from "react-dom";
 
-import { ImageTracker, ZapparCamera, ZapparCanvas, BrowserCompatibility, LogLevel, setLogLevel } from "../../../src/index";
+import { ImageTracker, ZapparCamera, ZapparCanvas, BrowserCompatibility, LogLevel, setLogLevel, Types, TargetImagePreviewMesh } from "../../../src/index";
 
 setLogLevel(LogLevel.LOG_LEVEL_VERBOSE);
 
-const App = () => {
+function App() {
   const imgSrc = require("file-loader!../../assets/sources/image-target.png").default;
   const target = require("file-loader!../../assets/target.zpt").default;
   const [loading, setLoading] = useState(true);
+  const ref = React.useRef<Types.ImageAnchorGroup>();
+  // eslint-disable-next-line no-undef
+  const [targetImagePreviewMesh, setTargetImagePreviewMesh] = useState<JSX.Element>();
 
   const img = useMemo(() => {
     const element = document.createElement("img");
@@ -28,7 +31,16 @@ const App = () => {
       <BrowserCompatibility fallback={<div>Sorry!</div>} />
       <ZapparCanvas>
         <ZapparCamera permissionRequest={false} sources={{ rearCamera: img }} />
-        <ImageTracker targetImage={target} onNewAnchor={() => console.log("Anchor is visible")}>
+        <ImageTracker
+          ref={ref}
+          targetImage={target}
+          onNewAnchor={() => {
+            const element = <TargetImagePreviewMesh imageTarget={ref.current!.imageTracker.targets[0]} />;
+            setTargetImagePreviewMesh(element);
+            console.log("Anchor is visible");
+          }}
+        >
+          {targetImagePreviewMesh}
           <mesh>
             <boxGeometry args={[1, 1, 1]} />
             <meshStandardMaterial color="orange" />
@@ -38,5 +50,5 @@ const App = () => {
       </ZapparCanvas>
     </>
   );
-};
+}
 render(<App />, document.getElementById("root"));