+++ tests/simple-image.test.ts
@@ -5,13 +5,13 @@ import * as util from "@zappar/jest-console-logs";
 import { toMatchImageSnapshot } from "jest-image-snapshot";
 
 expect.extend({ toMatchImageSnapshot });
-jest.setTimeout(60000);
+jest.setTimeout(60000 * 2);
 
 const url = "https://0.0.0.0:8082/pages/jest/module/simple-image.html";
 describe("image tracking", () => {
   it("console logs", async () => {
     const page = await browser.newPage();
-    page.goto(url);
+    page.goto(url, { timeout: 0 });
     await util.expectLogs({
       expected: [
         /^Zappar for React Three v/,
@@ -28,7 +28,7 @@ describe("image tracking", () => {
         "Anchor is visible",
       ],
       page,
-      timeoutMs: 30000,
+      timeoutMs: 60000,
     });
 
     // Wait 1s for texture to load